requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
pyTelegramBotAPI==4.4.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot-calendar==1.0.5
requests==2.27.1
six==1.16.0
urllib3==1.26.8
